+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master
 
 * pass Model#firebasePath function result through path expansion
+* add Model#firebasePathOptions to more easily change options passed to firebasePath
 
 ## v0.2.1